// src/components/game/generated/office-tmx-data.ts
export const ANIMATED_TILE_DEFS: Array<{ baseGid: number; frames: number[]; frameDurationSec: number }> = [
  { baseGid: 16301, frames: [16301,16302,16303,16304,16305,16306,16307,16308,16309,16310], frameDurationSec: 0.2 },
  { baseGid: 16321, frames: [16331,16332,16333,16334,16335,16336,16337,16338,16339,16340], frameDurationSec: 0.2 },
  { baseGid: 16341, frames: [16341,16342,16343,16344,16345,16346,16347,16348,16349,16350,16351,16352,16353,16354,16355,16356], frameDurationSec: 2 },
  { baseGid: 16357, frames: [16357,16358,16359,16360,16361,16362,16363,16364,16365,16366,16367,16368,16369,16370,16371,16372], frameDurationSec: 2 },
  { baseGid: 16373, frames: [16373,16374,16375,16376,16377,16378,16379,16380,16381,16382,16383,16384,16385,16386,16387,16388], frameDurationSec: 2 },
  { baseGid: 16437, frames: [16437,16438,16439,16440,16441], frameDurationSec: 0.2 },
  { baseGid: 16442, frames: [16442,16443,16444,16445,16446], frameDurationSec: 0.2 },
  { baseGid: 16543, frames: [16543,16544,16545], frameDurationSec: 0.2 },
  { baseGid: 16546, frames: [16546,16547,16548], frameDurationSec: 0.2 },
  { baseGid: 16549, frames: [16549,16550,16551], frameDurationSec: 0.2 },
]